Updated file names and component names for better organization by changing 'cards' to 'gameBoard' to differentiate between 'Cards' and 'Card' for others. #3
