+++ src/app.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import Cards from "./components/cards"
+import GameBoard from "./components/game-board"
 
 export default function App(){
   return (
@@ -8,7 +8,7 @@ export default function App(){
         <h1>Memory Challenge</h1>
       </div>
       <div className='row'>
-          { Cards() }
+          { GameBoard() }
       </div>
     </>
   )
